requirements.txt
fastapi
uvicorn
aiohttp
aiomysql
mysql-connector-python
python-dotenv
